# backend/src/api/auth.py
from fastapi import APIRouter, HTTPException
from typing import Optional
from pydantic import BaseModel
from datetime import datetime

router = APIRouter()

# Pydantic models for auth
class UserRegistration(BaseModel):
    email: str
    password: str
    name: str
    software_background: Optional[str] = "none"
    hardware_background: Optional[str] = "none"

class UserLogin(BaseModel):
    email: str
    password: str

class UserProfile(BaseModel):
    id: str
    email: str
    name: str
    software_background: str
    hardware_background: str
    created_at: datetime
    updated_at: datetime

# Mock implementation - would connect to Better Auth in real implementation
@router.post("/register", response_model=UserProfile)
async def register_user(user_data: UserRegistration):
    # This would integrate with Better Auth in the real implementation
    raise HTTPException(status_code=501, detail="Better Auth integration not yet implemented")

@router.post("/login")
async def login_user(credentials: UserLogin):
    # This would integrate with Better Auth in the real implementation
    raise HTTPException(status_code=501, detail="Better Auth integration not yet implemented")

@router.get("/profile", response_model=UserProfile)
async def get_profile():
    # This would integrate with Better Auth in the real implementation
    raise HTTPException(status_code=501, detail="Better Auth integration not yet implemented")