# backend/src/api/chapters.py
from fastapi import APIRouter, HTTPException, Query
from typing import List, Optional
from pydantic import BaseModel
from datetime import datetime

router = APIRouter()

# Pydantic models for chapters
class Chapter(BaseModel):
    id: str
    module_id: str
    chapter_number: int
    title: str
    slug: str
    content: str
    language: str
    created_at: datetime
    updated_at: datetime

class ChapterListResponse(BaseModel):
    chapters: List[Chapter]

class UpdateProgress(BaseModel):
    completion_percentage: int
    time_spent_seconds: Optional[int] = 0
    bookmarks: Optional[List[dict]] = []
    notes: Optional[List[dict]] = []

class UserProgress(BaseModel):
    id: str
    user_id: str
    chapter_id: str
    completion_percentage: int
    time_spent_seconds: int
    last_accessed: datetime
    bookmarks: List[dict]
    notes: List[dict]

@router.get("/", response_model=ChapterListResponse)
async def list_chapters(
    module_id: str,
    language: str = Query(default="en", description="Language for content")
):
    # This would fetch chapters from the database
    raise HTTPException(status_code=501, detail="Chapter listing not yet implemented")

@router.get("/{chapter_number}", response_model=Chapter)
async def get_chapter(
    module_id: str,
    chapter_number: int,
    language: str = Query(default="en", description="Language for content")
):
    # This would fetch a specific chapter from the database
    raise HTTPException(status_code=501, detail="Chapter retrieval not yet implemented")

@router.get("/{chapter_id}/progress", response_model=UserProgress)
async def get_user_progress(chapter_id: str):
    # This would fetch user progress for a chapter
    raise HTTPException(status_code=501, detail="User progress retrieval not yet implemented")

@router.post("/{chapter_id}/progress", response_model=UserProgress)
async def update_user_progress(chapter_id: str, progress: UpdateProgress):
    # This would update user progress for a chapter
    raise HTTPException(status_code=501, detail="User progress update not yet implemented")